requirements.txt
discord.py==2.3.0
nltk==3.8.1
numpy==1.23.4
numpy==1.24.3
tensorflow==2.11.0
tensorflow_intel==2.11.0
tflearn==0.5.0
